Remove the update debounce from the weekly-slot inside multislider
There was a bug in which the mergeOverlaps function would not merge newly created overlaps unless you waited for the debounce to finish before releasing the mouse button.

This solves that & makes the ui seem more responsive when dragging or resizing a slot around. This isn't that performance intensive, no real need to debounce.